Add color hooks for the mobile implementation.
