+++ packages/block-editor/src/hooks/color-panel.native.js
@@ -1,0 +1,3 @@
+export default function ColorPanel() {
+	return null;
+}
