+++ packages/block-editor/src/hooks/index.native.js
@@ -3,3 +3,5 @@
  */
 import './custom-class-name';
 import './generated-class-name';
+import './style';
+import './color';
