+++ packages/block-library/src/paragraph/edit.native.js
@@ -7,7 +7,6 @@ import {
 	AlignmentToolbar,
 	BlockControls,
 	RichText,
-	__experimentalUseColors,
 } from '@wordpress/block-editor';
 
 const name = 'core/paragraph';
@@ -21,12 +20,6 @@ function ParagraphBlock( {
 } ) {
 	const { align, content, placeholder } = attributes;
 
-	/* eslint-disable @wordpress/no-unused-vars-before-return */
-	const { TextColor } = __experimentalUseColors( [
-		{ name: 'textColor', property: 'color' },
-	] );
-	/* eslint-enable @wordpress/no-unused-vars-before-return */
-
 	return (
 		<>
 			<BlockControls>
@@ -37,35 +30,33 @@ function ParagraphBlock( {
 					} }
 				/>
 			</BlockControls>
-			<TextColor>
-				<RichText
-					identifier="content"
-					tagName="p"
-					value={ content }
-					deleteEnter={ true }
-					style={ style }
-					onChange={ ( nextContent ) => {
-						setAttributes( {
-							content: nextContent,
-						} );
-					} }
-					onSplit={ ( value ) => {
-						if ( ! value ) {
-							return createBlock( name );
-						}
+			<RichText
+				identifier="content"
+				tagName="p"
+				value={ content }
+				deleteEnter={ true }
+				style={ style }
+				onChange={ ( nextContent ) => {
+					setAttributes( {
+						content: nextContent,
+					} );
+				} }
+				onSplit={ ( value ) => {
+					if ( ! value ) {
+						return createBlock( name );
+					}
 
-						return createBlock( name, {
-							...attributes,
-							content: value,
-						} );
-					} }
-					onMerge={ mergeBlocks }
-					onReplace={ onReplace }
-					onRemove={ onReplace ? () => onReplace( [] ) : undefined }
-					placeholder={ placeholder || __( 'Start writing…' ) }
-					textAlign={ align }
-				/>
-			</TextColor>
+					return createBlock( name, {
+						...attributes,
+						content: value,
+					} );
+				} }
+				onMerge={ mergeBlocks }
+				onReplace={ onReplace }
+				onRemove={ onReplace ? () => onReplace( [] ) : undefined }
+				placeholder={ placeholder || __( 'Start writing…' ) }
+				textAlign={ align }
+			/>
 		</>
 	);
 }
